Improve error handling messages
For the SKABaseDevice, the logger may have failed to initialise so
we just print an error to stdout.  We don't want to raise another
exception in `init_device`.

Log exceptions on command failure to ensure we don't miss these.

The `!s` formatting for enums gives the enum name rather than the
integer value.  The name is easier to interpret.